+++ Cargo.toml
@@ -8,8 +8,7 @@ edition = "2021"
 [dependencies]
 iced = { git = "https://github.com/iced-rs/iced.git" , features = ["tokio", "advanced"] }
 tokio = {version = "1.21.2", features = ["sync"] }
-rustube = "0.6.0"
 image = "0.25.1"
-serde_json = "1"
 rusqlite = { version = "0.29.0", features = ["bundled"] }
-rodio = { version = "0.17.3", features = ["symphonia-all"] }
+rodio = { version = "0.17.3", features = ["symphonia-mp3"] }
+rusty_ytdl = "0.7.1"
