+++ src/ui/components/control_bar.rs
